config: T2485: move confdict.py to conf/interface.py
