+++ src/conf_mode/interfaces-wireless.py
@@ -24,7 +24,7 @@ from netifaces import interfaces
 from netaddr import EUI, mac_unix_expanded
 
 from vyos.config import Config
-from vyos.configdict import list_diff, intf_to_dict, add_to_dict, interface_default_data
+from vyos.config.interface import list_diff, intf_to_dict, add_to_dict, interface_default_data
 from vyos.ifconfig import WiFiIf, Section
 from vyos.ifconfig_vlan import apply_all_vlans, verify_vlan_config
 from vyos.template import render
